+++ rust/src/frb_generated.rs
@@ -39,7 +39,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.13.0";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -2138114040;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -131201380;
 
 // Section: executor
 
@@ -1720,6 +1720,41 @@ fn wire__crate__api__set_log_path_impl(
         },
     )
 }
+fn wire__crate__api__set_playback_speed_impl(
+    port_: flutter_rust_bridge::for_generated::MessagePort,
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_normal::<flutter_rust_bridge::for_generated::SseCodec, _, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "set_playback_speed",
+            port: Some(port_),
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Normal,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_speed = <f32>::sse_decode(&mut deserializer);
+            deserializer.end();
+            move |context| {
+                transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
+                    (move || {
+                        let output_ok = crate::api::set_playback_speed(api_speed)?;
+                        std::result::Result::Ok(output_ok)
+                    })(),
+                )
+            }
+        },
+    )
+}
 fn wire__crate__api__set_replay_gain_impl(
     port_: flutter_rust_bridge::for_generated::MessagePort,
     ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
@@ -2230,12 +2265,13 @@ fn pde_ffi_dispatcher_primary_impl(
         46 => wire__crate__api__set_gapless_impl(port, ptr, rust_vec_len, data_len),
         47 => wire__crate__api__set_haas_ms_impl(port, ptr, rust_vec_len, data_len),
         48 => wire__crate__api__set_log_path_impl(port, ptr, rust_vec_len, data_len),
-        49 => wire__crate__api__set_replay_gain_impl(port, ptr, rust_vec_len, data_len),
-        50 => wire__crate__api__set_skip_silence_impl(port, ptr, rust_vec_len, data_len),
-        51 => wire__crate__api__set_soft_clip_impl(port, ptr, rust_vec_len, data_len),
-        52 => wire__crate__api__set_stereo_width_impl(port, ptr, rust_vec_len, data_len),
-        53 => wire__crate__api__set_volume_impl(port, ptr, rust_vec_len, data_len),
-        54 => wire__crate__api__stop_impl(port, ptr, rust_vec_len, data_len),
+        49 => wire__crate__api__set_playback_speed_impl(port, ptr, rust_vec_len, data_len),
+        50 => wire__crate__api__set_replay_gain_impl(port, ptr, rust_vec_len, data_len),
+        51 => wire__crate__api__set_skip_silence_impl(port, ptr, rust_vec_len, data_len),
+        52 => wire__crate__api__set_soft_clip_impl(port, ptr, rust_vec_len, data_len),
+        53 => wire__crate__api__set_stereo_width_impl(port, ptr, rust_vec_len, data_len),
+        54 => wire__crate__api__set_volume_impl(port, ptr, rust_vec_len, data_len),
+        55 => wire__crate__api__stop_impl(port, ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }